requirements.txt
matplotlib==3.8.4
pandas==2.2.2
plotly==5.22.0
rdkit_pypi==2022.9.5
SQLAlchemy==2.0.30
streamlit==1.35.0
